chore: restructure io manager to split out custom db io manager
